Extract UpdateSnapshotStatusAction to its own file
Pull a Transport*Action out of SnapshotsService
in order to simplify and focus the SnapshotsService
code.

Relates ES-11650
